// lib/wordpress.d.ts
interface WPEntity {
  id: number;
  date: string;
  date_gmt: string;
  modified: string;
  modified_gmt: string;
  slug: string;
  status: "publish" | "future" | "draft" | "pending" | "private";
  link: string;
  guid: {
    rendered: string;
  };
}

interface RenderedContent {
  rendered: string;
  protected: boolean;
}

interface RenderedTitle {
  rendered: string;
}

// Media types
interface MediaSize {
  file: string;
  width: number;
  height: number;
  mime_type: string;
  source_url: string;
}

interface MediaDetails {
  width: number;
  height: number;
  file: string;
  sizes: Record<string, MediaSize>;
}

export interface FeaturedMedia extends WPEntity {
  title: RenderedTitle;
  author: number;
  caption: {
    rendered: string;
  };
  alt_text: string;
  media_type: string;
  mime_type: string;
  media_details: MediaDetails;
  source_url: string;
}
// ...
export interface Project extends WPEntity {
  title: RenderedTitle;
  content: RenderedContent;
  excerpt: RenderedContent;
  author: number;
  featured_media: number;
  template: string;
  categories: number[];
  meta: {
    heading_one?: string;
    heading_two?: string;
    project_images?: number[];
    [key: string]: unknown;
  };
  _embedded?: PostEmbedded;
}
// ...
interface EmbeddedAuthor {
  id: number;
  name: string;
  slug: string;
  avatar_urls: Record<string, string>;
}

interface EmbeddedTerm {
  id: number;
  name: string;
  slug: string;
}

export interface PostEmbedded {
  author?: EmbeddedAuthor[];
  "wp:featuredmedia"?: FeaturedMedia[];
  "wp:term"?: EmbeddedTerm[][];
}